requirements.txt
asttokens==2.1.0
backcall==0.2.0
colorama==0.4.6
DateTime==5.5
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
filelock==3.15.4
fsspec==2024.6.1
ipykernel==6.17.1
ipython==8.6.0
jedi==0.18.1
Jinja2==3.1.4
jupyter_client==7.4.6
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.6
networkx==3.3
numpy==1.26.4
pandas==2.2.2
parso==0.8.3
pickleshare==0.7.5
platformdirs==2.5.4
prompt-toolkit==3.0.32
psutil==5.9.4
pure-eval==0.2.2
Pygments==2.13.0
python-box==7.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
pyzmq==24.0.1
scipy==1.14.0
six==1.16.0
stack-data==0.6.1
sympy==1.13.2
torch==2.4.0
tornado==6.2
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
wcwidth==0.2.5
zope.interface==7.0.1
